pyproject.toml
[project]
name = "todoist"
version = "0.1.0"
description = "FastAPI app with Auth0 and enroll endpoint"
requires-python = "~=3.10"
authors = [
    {name = "Tijmen van Graft", email = "tijmen.graft@helindata.com"}
]
classifiers = ["Private :: Do Not Upload"]
dependencies = [
    "fastapi >= 0.110.0",
    "uvicorn >= 0.29.0",
    "fastapi-auth0 >= 0.5.0",
    "requests >= 2.32.4",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "rich >= 13.7.1",
    "pre-commit (>=4.2.0,<5.0.0)",
    "ruff (>=0.12.5,<0.13.0)",
    "pyflakes (>=3.4.0,<4.0.0)",
    "starlette (>=0.47.2)",
    "authlib (>=1.6.5,<2.0.0)",
    "python-keycloak (>=5.8.1,<6.0.0)",
]

[tool.poetry]
requires-poetry = ">=2.0"
packages = [
    { include = "todoist", from = "./src" },
]

[tool.poetry.group.test.dependencies]
pytest = "^8.4.1"
pytest-cov = "^6.2.1"
pytest-asyncio = "^1.1.0"
httpx = "^0.28.1"
pytest-env = "^1.1.5"

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
asyncio_mode = "auto"
log_cli = true
log_cli_level = "DEBUG"
env = [
    "postgres_user=postgres",
    "postgres_password=helin_password_12312341fda",
    "postgres_db=helin",
    "postgres_server=postgres",
]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.alembic]

# path to migration scripts.
# this is typically a path given in POSIX (e.g. forward slashes)
# format, relative to the token %(here)s which refers to the location of this
# ini file
script_location = "src/alembic"

# template used to generate migration file names; The default value is %%(rev)s_%%(slug)s
# Uncomment the line below if you want the files to be prepended with date and time
# see https://alembic.sqlalchemy.org/en/latest/tutorial.html#editing-the-ini-file
# for all available tokens
# file_template = "%%(year)d_%%(month).2d_%%(day).2d_%%(hour).2d%%(minute).2d-%%(rev)s_%%(slug)s"

# additional paths to be prepended to sys.path. defaults to the current working directory.
prepend_sys_path = [
    "."
]

# timezone to use when rendering the date within the migration file
# as well as the filename.
# If specified, requires the python>=3.9 or backports.zoneinfo library and tzdata library.
# Any required deps can installed by adding `alembic[tz]` to the pip requirements
# string value is passed to ZoneInfo()
# leave blank for localtime
# timezone =

# max length of characters to apply to the "slug" field
# truncate_slug_length = 40

# set to 'true' to run the environment during
# the 'revision' command, regardless of autogenerate
# revision_environment = false

# set to 'true' to allow .pyc and .pyo files without
# a source .py file to be detected as revisions in the
# versions/ directory
# sourceless = false

# version location specification; This defaults
# to <script_location>/versions.  When using multiple version
# directories, initial revisions must be specified with --version-path.
# version_locations = [
#    "%(here)s/alembic/versions",
#    "%(here)s/foo/bar"
# ]


# set to 'true' to search source files recursively
# in each "version_locations" directory
# new in Alembic version 1.10
# recursive_version_locations = false

# the output encoding used when revision files
# are written from script.py.mako
# output_encoding = "utf-8"

# This section defines scripts or Python functions that are run
# on newly generated revision scripts.  See the documentation for further
# detail and examples
# [[tool.alembic.post_write_hooks]]
# format using "black" - use the console_scripts runner,
# against the "black" entrypoint
# name = "black"
# type = "console_scripts"
# entrypoint = "black"
# options = "-l 79 REVISION_SCRIPT_FILENAME"
#
# [[tool.alembic.post_write_hooks]]
# lint with attempts to fix using "ruff" - use the module runner, against the "ruff" module
# name = "ruff"
# type = "module"
# module = "ruff"
# options = "check --fix REVISION_SCRIPT_FILENAME"
#
# [[tool.alembic.post_write_hooks]]
# Alternatively, use the exec runner to execute a binary found on your PATH
# name = "ruff"
# type = "exec"
# executable = "ruff"
# options = "check --fix REVISION_SCRIPT_FILENAME"
